Agregar usuarios depende el rol

+++ database/migrations/2021_03_28_022204_create_pesos_table.php
@@ -16,10 +16,9 @@ class CreatePesosTable extends Migration
         Schema::create('pesos', function (Blueprint $table) {
             $table->id();
             $table->foreignId("id_semillas")->references("id")->on("semillas");
-            $table->float('PesoGrande');
-            $table->float('PesoMediano');
-            $table->float('PesoPequeno');
-
+            $table->decimal('PesoGrande',50,2);
+            $table->decimal('PesoMediano',50,2);
+            $table->decimal('PesoPequeno',50,2);
             $table->timestamps();
         });
     }
